utils/web_scraper.py
import requests
from bs4 import BeautifulSoup

def fetch_job_description(url):
    resp = requests.get(url, timeout=10)
    soup = BeautifulSoup(resp.text, 'html.parser')
    paragraphs = soup.find_all('p')
    text = "\n".join([p.get_text() for p in paragraphs])
    return text[:2000]