requirements.txt
streamlit
requests
beautifulsoup4
PyPDF2
openai
python-docx